Adding support for local Minions.
In order to provide unique and immutable IDs for Minions, a StatefulSet
was used, even if the Minions are stateless. A Deployment doesn't make
sense because the Pod names are dynamic and will change if the Pod
should be re-scheduled.

The README was updated to be more consistent, specially considering that
now the solution include Minions.

The scripts were improved.