Adding support for Kubeless.
The Kubeless deployment has been tested and it works as expected.
That said, the Fission deployment should be re-tested to make sure
it still works, as now, secrets are used instead of config-maps for
the slackUrl, and also the JS file doesn't have a default handler,
and you have to speficy an entrypoint.