Deprecate kafka-converter in favor of producer-enhancer
The serverless example uses the JSON payload generated from the protobuf
data that the OpenNMS Kafka Produder produces.

Then, a KSQL application was added to show how to enhance an alarm with
node data when available.

The above was simplified with the addition of the producer-enhancer
which is a Kafka Stream application implemented with Goka to produce a
JSON payload that contains the alarm and also the node data when
available on the same message.

With this in place, the KSQL example is obsolete, and the serverless
implementation can use the new playload.