event-watcher: don't parse resource version
Parsing the resourceVersion as an integer is strictly prohibited by the
Kubernetes API [1]. During a review of code that would be affected by
non-numerical resourceVersions, I came upon this code.

It looks like this code was using resource version comparison to only
handle "new" objects. The Kubernetes watch API ensures that events on a
watch stream are delivered once and only once, and always in order.
Therefore, no check of this sort is necessary as each new event will
always be unique and newer than any previous event.

[1] https://kubernetes.io/docs/reference/using-api/api-concepts/#resource-versions

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>